requirements_phase2.txt
# Phase 2 - Vision System Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
mss>=9.0.0
Pillow>=10.0.0
requests>=2.31.0
